% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mmc.R
\name{fit.mmc}
\alias{fit.mmc}
\title{performance measures}
\usage{
fit.mmc(data, n, c, alpha)
}
\arguments{
\item{data}{data set.}

\item{n}{number of simulations interest you want to perform.}

\item{c}{number of servers.}

\item{alpha}{significance level for the Kolmogorov-Smirnov test.}
}
\value{
a list of all performance measures used.
}
\description{
It allows to evaluate the efficiency of a system by analyzing its characteristics using performance measures.
}
\author{
Mayara Almeida
}
