% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.mm1.R
\name{r.mm1}
\alias{r.mm1}
\title{data simulation for the model M\ M\ 1}
\usage{
r.mm1(n, lambda, mu)
}
\arguments{
\item{n}{number of simulations interest you want to perform.}

\item{lambda}{average rate that represents an exponential distribution for arrival times.}

\item{mu}{average rate that represents an exponential distribution for the times of services.}
}
\value{
will return values in the form of  three column matrix, with the arrival, service and departure times.
}
\description{
function for the simulation of data imitating the operation of a queue model.
}
\author{
Mayara Almeida
}
