% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.MMC_pacience.R
\name{r.MMC_pacience}
\alias{r.MMC_pacience}
\title{data simulation for the model M\ M\ c with withdrawal.}
\usage{
r.MMC_pacience(n, c, lambda, mu, tau)
}
\arguments{
\item{n}{number of simulations interest you want to perform.}

\item{lambda}{average rate that represents an exponential distribution for arrival times.}

\item{mu}{average rate that represents an exponential distribution for the times of services.}

\item{tau}{average rate that represents an exponential distribution for the times of tolerance}
}
\value{
will return values in the form of  three column matrix, with the arrival, service and departure times.
}
\description{
function for the simulation of data imitating the operation of a queue model considering withdrawal.
}
\author{
Debora de Sousa Cordeiro
}
