% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.queue.R
\name{r.queue}
\alias{r.queue}
\title{function for data simulation}
\usage{
r.queue(n, c, lambda, mu)
}
\arguments{
\item{n}{number of simulations interest you want to perform.}

\item{c}{number of servers.}

\item{lambda}{average rate that represents an exponential distribution for arrival times.}

\item{mu}{average rate that represents an exponential distribution for the times of services.}
}
\value{
will return values in the form of a matrix of four columns, with the arrival times, the times when the services were started, the departures and the number of the counter that was attended.
}
\description{
function for the simulation of data imitating the operation of a queue model.
}
\examples{
set.seed(10)
data <- r.queue(10,2,1,4)
}
\author{
Mayara Almeida
}
