% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqt.mm1.R
\name{wqt.mm1}
\alias{wqt.mm1}
\alias{Wqt.mm1}
\title{Cumulative distribution function}
\usage{
Wqt.mm1(t, lambda, mu)
}
\arguments{
\item{t}{waiting time.}

\item{lambda}{average rate that represents an exponential distribution for arrival times.}

\item{mu}{average rate that represents an exponential distribution for the times of services.}
}
\value{
will return the probability value of the time that any user remains in the queue waiting for assistance.
}
\description{
Cumulative queue waiting time distribution function.
}
\examples{
Wqt.mm1(5,3,4)
}
\author{
Mayara Almeida
}
