% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqt.mmc.R
\name{wqt.mmc}
\alias{wqt.mmc}
\alias{Wqt.mmc}
\title{Cumulative distribution function}
\usage{
Wqt.mmc(t, c, lambda, mu)
}
\arguments{
\item{t}{waiting time.}

\item{c}{number of servers.}

\item{lambda}{average rate that represents an exponential distribution for arrival times.}

\item{mu}{average rate that represents an exponential distribution for the times of services.}
}
\value{
will return the probability value of the time that any user remains in the queue waiting for assistance.
}
\description{
Cumulative queue waiting time distribution function.
}
\author{
Mayara Almeida
}
