% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt.mm1.R
\name{wt.mm1}
\alias{wt.mm1}
\alias{Wt.mm1}
\title{Cumulative distribution function}
\usage{
Wt.mm1(t, lambda, mu)
}
\arguments{
\item{t}{waiting time.}

\item{lambda}{average rate that represents an exponential distribution for arrival times.}

\item{mu}{average rate that represents an exponential distribution for the times of services.}
}
\value{
will return the probability value of the time that any user remains on the system.
}
\description{
Cumulative waiting time distribution function
}
\examples{
 Wt.mm1<- function(5,3, 4)
}
\author{
Mayara Almeida
}
