% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mm1.R
\name{fit.mm1}
\alias{fit.mm1}
\title{performance measures}
\usage{
fit.mm1(n, data, alpha)
}
\arguments{
\item{n}{number of simulations interest you want to perform.}

\item{data}{data set.}

\item{alpha}{significance level for the Kolmogorov-Smirnov test.}
}
\value{
a list of all performance measures used.
}
\description{
It allows to evaluate the efficiency of a system by analyzing its characteristics using performance measures.
}
\author{
Mayara Almeida
}
